+++ demo/reagentdemo/intro.cljs
@@ -1,5 +1,6 @@
 (ns reagentdemo.intro
   (:require [reagent.core :as r]
+            [reagent.dom :as rdom]
             [reagentdemo.syntax :as s]
             [reagentdemo.common :as common :refer [demo-component]]
             [simpleexample.core :as simple]
@@ -62,7 +63,8 @@
        "Seconds Elapsed: " @seconds-elapsed])))
 
 (defn render-simple []
-  (r/render [simple-component]
+  (rdom/render
+    [simple-component]
     (.-body js/document)))
 
 (def bmi-data (r/atom {:height 180 :weight 80}))
@@ -222,7 +224,7 @@
 
    [:p "Reagent supports most of React’s API, but there is really only
    one entry-point that is necessary for most applications: "
-    [:code "reagent.core/render"] "."]
+    [:code "reagent.dom/render"] "."]
 
    [:p "It takes two arguments: a component, and a DOM node. For
    example, splashing the very first example all over the page would