+++ src/reagent/core.cljs
@@ -9,8 +9,7 @@
             [reagent.ratom :as ratom]
             [reagent.debug :as deb :refer-macros [assert-some assert-component
                                                   assert-js-object assert-new-state
-                                                  assert-callable]]
-            [reagent.dom :as dom]))
+                                                  assert-callable]]))
 
 (def is-client util/is-client)
 
@@ -63,43 +62,6 @@
   [c]
   (assert-some c "Component")
   (comp/reactify-component c))
-
-(defn render
-  "Render a Reagent component into the DOM. The first argument may be
-  either a vector (using Reagent's Hiccup syntax), or a React element.
-  The second argument should be a DOM node.
-
-  Optionally takes a callback that is called when the component is in place.
-
-  Returns the mounted component instance."
-  ([comp container]
-   (dom/render comp container))
-  ([comp container callback]
-   (dom/render comp container callback)))
-
-(defn unmount-component-at-node
-  "Remove a component from the given DOM node."
-  [container]
-  (dom/unmount-component-at-node container))
-
-;; For backward compatibility
-(def ^{:deprecated "0.10.0"} as-component as-element)
-(def ^{:deprecated "0.10.0"} render-component render)
-
-(defn force-update-all
-  "Force re-rendering of all mounted Reagent components. This is
-  probably only useful in a development environment, when you want to
-  update components in response to some dynamic changes to code.
-
-  Note that force-update-all may not update root components. This
-  happens if a component 'foo' is mounted with `(render [foo])` (since
-  functions are passed by value, and not by reference, in
-  ClojureScript). To get around this you'll have to introduce a layer
-  of indirection, for example by using `(render [#'foo])` instead."
-  []
-  (ratom/flush!)
-  (dom/force-update-all)
-  (batch/flush-after-render))
 
 (defn create-class
   "Creates JS class based on provided Clojure map, for example:
@@ -210,11 +172,6 @@
   (assert-component this)
   (comp/get-argv this))
 
-(defn dom-node
-  "Returns the root DOM node of a mounted component."
-  [this]
-  (dom/dom-node this))
-
 (defn class-names
   "Function which normalizes and combines class values to a string
 
@@ -238,7 +195,7 @@
   ([defaults props & others] (apply util/merge-props defaults props others)))
 
 (defn flush
-  "Render dirty components immediately to the DOM.
+  "Render dirty components immediately.
 
   Note that this may not work in event handlers, since React.js does
   batching of updates there."
@@ -393,6 +350,6 @@
   "Try to return the path of component c as a string.
   Maybe useful for debugging and error reporting, but may break
   with future versions of React (and return nil)."
-  {:deprecated "0.10.0"}
+  {:deprecated "1.0.0"}
   [c]
   (comp/component-path c))