+++ test-environments/browser-cljsjs-prod/test.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 set -ex
-lein do clean, doo chrome-headless prod-test once
+rm -rf target/cljsbuild/prod-test/
+lein doo chrome-headless prod-test once
 test -f target/cljsbuild/prod-test/main.js
 node_modules/.bin/gzip-size target/cljsbuild/prod-test/main.js