+++ test-environments/browser-cljsjs/test.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
 set -ex
-lein do clean, doo chrome-headless test once
+rm -rf target/cljsbuild/test/
+lein doo chrome-headless test once
 test -f target/cljsbuild/test/out/cljsjs/react/development/react.inc.js