+++ test-environments/browser-npm-prod/test.sh
@@ -2,6 +2,7 @@
 
 set -ex
 
-lein do clean, doo chrome-headless prod-test-npm once
+rm -rf target/cljsbuild/prod-test-npm/
+lein doo chrome-headless prod-test-npm once
 test -f target/cljsbuild/prod-test-npm/main.js
 node_modules/.bin/gzip-size target/cljsbuild/prod-test-npm/main.js