+++ test-environments/browser-npm/test.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
 set -ex
-lein do clean, doo chrome-headless test-npm once
+rm -rf target/cljsbuild/test-npm/
+lein doo chrome-headless test-npm once
 test -f target/cljsbuild/test-npm/out/node_modules/react/index.js