+++ test-environments/node-cljsjs/test.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
 set -ex
-lein do clean, doo node node-test once
+rm -rf target/cljsbuild/node-test/
+lein doo node node-test once
 test -f target/cljsbuild/node-test/out/cljsjs/react/development/react.inc.js