+++ test/reagenttest/utils.cljs
@@ -1,22 +1,23 @@
 (ns reagenttest.utils
-  (:require [reagent.core :as r]))
+  (:require [reagent.core :as r]
+            [reagent.dom :as rdom]))
 
 (defn with-mounted-component [comp f]
   (when r/is-client
     (let [div (.createElement js/document "div")]
       (try
-        (let [c (r/render comp div)]
+        (let [c (rdom/render comp div)]
           (f c div))
         (finally
-          (r/unmount-component-at-node div)
+          (rdom/unmount-component-at-node div)
           (r/flush))))))
 
 (defn with-mounted-component-async [comp done f]
   (when r/is-client
     (let [div (.createElement js/document "div")
-          c (r/render comp div)]
+          c (rdom/render comp div)]
       (f c div (fn []
-                 (r/unmount-component-at-node div)
+                 (rdom/unmount-component-at-node div)
                  (r/flush)
                  (done))))))